Initial Markov Chain experiment
Quick and dirty first dive into making a Markov Chain from scratch.
Current implementation takes in a text file and then puts all of the
words of that file into a word pair dictionary who's keys are two
consecutive words and who's values are a list of single words that
follow that word pair. When creating the chain, the function will
randomly pick one of the following words from the dictionary, and use
the new last two words to search for another word until a '[$STOP]' is
hit.

TODO:
- Add Comments to make it all more readable
- Refactor the creation of the chain out of __init__()
- Go through and verify variable names to make sure they are relevant to
their function